fix(database): review fixes
