+++ database/README.md
@@ -1,12 +1,12 @@
 # @mse/database
 
-Пакет для работы с Prisma ORM и локальной MySQL (без Docker).
+Пакет с Prisma Client для работы с MySQL.
 
 ## Требования
 
 - Node.js 20+
 - pnpm 10+
-- Локальная MySQL (или MariaDB) на `localhost:3306`
+- Доступная MySQL (локально или в Docker)
 
 ## Настройка
 
@@ -20,16 +20,24 @@ pnpm install
 
 ```env
 DATABASE_URL="mysql://root:your_password@localhost:3306/mse_db"
-DB_USER=root
-DB_PASSWORD=your_password
-DB_NAME=mse_db
 ```
 
-1. Создать локальную базу:
+Для Docker можно использовать сервис БД как host в URL, например:
+
+```env
+DATABASE_URL="mysql://root:your_password@db:3306/mse_db"
+```
+
+3. Если база запускается локально, создать БД:
 
 ```bash
 mysql -u root -p -e "CREATE DATABASE IF NOT EXISTS mse_db CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci;"
 ```
+
+При запуске пакет проверяет `DATABASE_URL`:
+- переменная обязательна;
+- URL должен быть валидным;
+- протокол только `mysql://` или `mariadb://`.
 
 ## Команды Prisma
 
@@ -51,13 +59,27 @@ pnpm d prisma:push
 pnpm d prisma:migrate -- --name init
 ```
 
-## Запуск тестового скрипта
+## Использование клиента в backend
+
+Пакет экспортирует `prisma`-клиент и типы Prisma:
+
+```ts
+import { prisma } from '@mse/database';
+```
+
+## hello_world-скрипт
+
+```bash
+pnpm d hello_world
+```
+
+Также доступно:
 
 ```bash
 pnpm d dev
 ```
 
-Скрипт `database/src/index.ts` выполняет простой `create` и `findMany`.
+Скрипт `database/src/scripts/hello_world.ts` удаляет тестового пользователя `test@example.com`, создаёт его заново и выводит список пользователей.
 
 ## Проверка кода
 
@@ -67,4 +89,3 @@ pnpm d lint
 pnpm d format
 pnpm d format:check
 ```
-
