+++ database/src/index.ts
@@ -1,36 +1,9 @@
-import "dotenv/config";
 import { PrismaMariaDb } from "@prisma/adapter-mariadb";
 import { PrismaClient } from '../generated/prisma/client';
+import { getDatabaseUrl } from './env';
 
-const adapter = new PrismaMariaDb({
-  host: "localhost",
-  port: 3306,
-  user: process.env.DB_USER,
-  password: process.env.DB_PASSWORD,
-  database: process.env.DB_NAME,
-});
+const adapter = new PrismaMariaDb(getDatabaseUrl());
 
-const prisma = new PrismaClient({ adapter });
+export const prisma = new PrismaClient({ adapter });
 
-async function main(): Promise<void> {
-  const createdUser = await prisma.user.create({
-    data: {
-      email: 'test@example.com',
-      name: 'Test User',
-    },
-  });
-
-  console.log('Created user:', createdUser);
-
-  const users = await prisma.user.findMany();
-  console.log('All users:', users);
-}
-
-main()
-  .catch((error: unknown) => {
-    console.error(error);
-    process.exit(1);
-  })
-  .finally(async () => {
-    await prisma.$disconnect();
-  });
+export * from '../generated/prisma/client';
